Change module organization
