+++ docs/api/code.rst
@@ -1,7 +1,8 @@
-Module *evaluator*
-==================
+Module *code*
+=============
 
 .. automodule:: sismic.code
     :members:
     :member-order: bysource
+    :show-inheritance:
