+++ docs/api/exceptions.rst
@@ -4,4 +4,4 @@ Module *exceptions*
 .. automodule:: sismic.exceptions
     :members:
     :member-order: bysource
-
+    :show-inheritance:
