+++ docs/api/interpreter.rst
@@ -4,4 +4,5 @@ Module *interpreter*
 .. automodule:: sismic.interpreter
     :members:
     :member-order: bysource
+    :show-inheritance:
