+++ docs/api/io.rst
@@ -4,4 +4,5 @@ Module *io*
 .. automodule:: sismic.io
     :members:
     :member-order: bysource
+    :show-inheritance:
