+++ docs/api/model.rst
@@ -4,4 +4,5 @@ Module *model*
 .. automodule:: sismic.model
     :members:
     :member-order: bysource
+    :show-inheritance:
