+++ docs/api/stories.rst
@@ -4,4 +4,5 @@ Module *stories*
 .. automodule:: sismic.stories
     :members:
     :member-order: bysource
+    :show-inheritance:
