+++ docs/api/testing.rst
@@ -4,4 +4,5 @@ Module *testing*
 .. automodule:: sismic.testing
     :members:
     :member-order: bysource
+    :show-inheritance:
