+++ sismic/code/__init__.py
@@ -1,0 +1,5 @@
+from .evaluator import Evaluator
+from .dummy import DummyEvaluator
+from .python import PythonEvaluator
+
+__all__ = ['Evaluator', 'DummyEvaluator', 'PythonEvaluator']
